chore: renamed signIn to sign-in to better fit naming conventions
