+++ packages/react/src/hooks/index.ts
@@ -4,9 +4,9 @@ export * from "./client.js";
 export * from "./latest-blockhash.js";
 export * from "./program-accounts.js";
 export * from "./recent-prioritization-fees.js";
+export * from "./sign-in.js";
 export * from "./signature-statuses.js";
 export * from "./signatures-for-address.js";
-export * from "./signIn.js";
 export * from "./slot.js";
 export * from "./token-account.js";
 export * from "./token-mint.js";
